new exceptions with handling

+++ GreetingService.Infrastructure/UserServices/SqlUserService.cs
@@ -1,4 +1,6 @@
-﻿using GreetingService.Core.Entities;
+﻿using GreetingService.Core.Extensions;
+using GreetingService.Core.Entities;
+using GreetingService.Core.Exceptions;
 using GreetingService.Core.Interfaces;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.Logging;
@@ -29,7 +31,7 @@ public class SqlUserService : IUserServiceAsync
         else
         {
             _logger.LogWarning($"User '{email}' not found.");
-            return false;
+            throw new UserDoesNotExistException();
         }
     }
 
@@ -37,12 +39,21 @@ public class SqlUserService : IUserServiceAsync
     {
         try
         {
+            User? existingUser = await _db.Users.FindAsync(user.Email);
+            if (existingUser != null)
+                throw new UserAlreadyExistsException(user.Email);
+
             await _db.Users.AddAsync(user);
             await _db.SaveChangesAsync();
             return true;
         }
+        catch (Exception e) when (e is UserAlreadyExistsException or InvalidEmailException)
+        {
+            throw;
+        }
         catch (Exception ex)
         {
+            _logger.LogError(ex.ToString());
             return false;
         }
     }
@@ -50,7 +61,11 @@ public class SqlUserService : IUserServiceAsync
     public async Task<User> GetUserAsync(string email)
     {
         User? user = await _db.Users.FindAsync(email);
-        return user;
+
+        if (user != null)
+            return user;
+        else
+            throw new UserDoesNotExistException(email);
     }
 
     public async Task<bool> UpdateUserAsync(User updatedUser)
@@ -65,23 +80,26 @@ public class SqlUserService : IUserServiceAsync
         }
         else
         {
-            return false;
+            throw new UserDoesNotExistException(updatedUser.Email);
         }
     }
 
     public async Task<bool> DeleteUserAsync(string email)
     {
+        if (!email.IsValidEmailAddress())
+            throw new InvalidEmailException(email);
+
         User? user = await _db.Users.FindAsync(email);
 
         if (user != null)
         {
             await Task.Run(() => _db.Users.Remove(user));
-            await _db.SaveChangesAsync();   
+            await _db.SaveChangesAsync();
             return true;
         }
         else
         {
-            return false;
+            throw new UserDoesNotExistException(email);
         }
     }
 }
